==> 3. Source_Sort.ipynb <==
import numpy as np
import os
import pandas as pd

#Cluster directories:
M35_dir = 'M35'
M67_dir = 'M67'
M67v2_dir = 'M67v2'
NGC869_dir = 'NGC869'

#Setting Cluster to process
cluster_dir = NGC869_dir  #------------------------------------------!!!!!

#Output directories to shared and local folders
shared_dir = os.path.join('/data/observatory/student_data/William_ODonnell', cluster_dir)
local_dir = cluster_dir

Gdf = pd.read_csv(os.path.join(local_dir, "2. "+ cluster_dir + '_G_data.csv'))
Idf = pd.read_csv(os.path.join(local_dir, "2. "+ cluster_dir + '_I_data.csv'))
Rdf = pd.read_csv(os.path.join(local_dir, "2. "+ cluster_dir + '_R_data.csv'))

ID = (Gdf['ID'])
RA = (Gdf['Right-Ascension'])
DEC = (Gdf['Declination'])
G_mag = (Gdf['G_mag'])
I_mag = (Idf['I_mag'])
R_mag = (Rdf['R_mag'])

Final = np.vstack((ID, RA, DEC, G_mag, I_mag, R_mag)).T
Final_df = pd.DataFrame(Final, columns = ['ID', 'Right-Ascension', 'Declination', 'G_Mag','I_Mag', 'R_Mag'], dtype = np.float64)
Final_df = Final_df.round({'Right-Ascension': 3, 'Declination': 3})
Final_df = Final_df.sort_values(by = ['Right-Ascension', 'Declination'], ascending = False )
Final_df.to_csv(os.path.join(local_dir, "3. " + cluster_dir + "_Uncalibrated_mags.csv"))
